[ci] Add Dependabot configuration
Added a Dependabot configuration file `dependabot.yml` at the root directory of the project. The configuration specifies the package ecosystem as "maven" and sets the update schedule to run weekly. Dependabot will now automatically check for and notify about any outdated Maven dependencies and propose updates, contributing to a more secure and up-to-date project.