[deps] Add Tomcat dependency to pom.xml and comment it out
This commit adds the Tomcat dependency to the `pom.xml` file, allowing for Tomcat to be used as the application server for this project. However, the newly added dependency is immediately commented out, preventing its active usage.

The purpose of adding the dependency and commenting it out is to have Tomcat available as an option for deployment or testing in the future, but not to include it in the current build. This decision was made to maintain flexibility and ease of switching between different application servers based on project requirements.

By using this approach, the project retains the option to enable Tomcat easily when needed without affecting the current build configuration.